:recycle: Refactor init file
