+++ src/main.rs
@@ -6,10 +6,7 @@ use crate::error::MocksError;
 use crate::server::Server;
 use crate::storage::Storage;
 use clap::Parser;
-use std::fs;
-use std::io::{self, Write};
 use std::net::{IpAddr, SocketAddr};
-use std::path::Path;
 
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
@@ -75,7 +72,7 @@ async fn main() -> Result<(), MocksError> {
             Server::startup(socket_addr, &url, storage).await?;
         }
         Commands::Init(args) => {
-            init_storage_file(&args.file, args.empty)?;
+            Storage::init_file(&args.file, args.empty)?;
         }
     }
 
@@ -95,57 +92,6 @@ fn init(host: &str, port: u16) -> Result<SocketAddr, MocksError> {
         .map_err(|e| MocksError::InvalidArgs(e.to_string()))
 }
 
-fn init_storage_file(file_path: &str, empty: bool) -> Result<(), MocksError> {
-    let path = Path::new(file_path);
-    
-    if path.exists() {
-        print!("File {} already exists. Overwrite? (y/N): ", file_path);
-        io::stdout().flush().unwrap();
-        
-        let mut input = String::new();
-        io::stdin().read_line(&mut input).unwrap();
-        
-        if !input.trim().to_lowercase().starts_with('y') {
-            println!("Aborted.");
-            return Ok(());
-        }
-    }
-    
-    if let Some(parent) = path.parent() {
-        if !parent.exists() {
-            fs::create_dir_all(parent)
-                .map_err(|e| MocksError::InvalidArgs(format!("Failed to create directory: {}", e)))?;
-        }
-    }
-    
-    let content = if empty {
-        r#"{
-  "posts": [],
-  "profile": {}
-}"#
-    } else {
-        r#"{
-  "posts": [
-    {
-      "id": 1,
-      "title": "Hello World",
-      "content": "This is a sample post"
-    }
-  ],
-  "profile": {
-    "id": 1,
-    "name": "Sample User"
-  }
-}"#
-    };
-    
-    fs::write(path, content)
-        .map_err(|e| MocksError::InvalidArgs(format!("Failed to write file: {}", e)))?;
-    
-    println!("Created: {}", file_path);
-    Ok(())
-}
-
 fn print_startup_info(url: &str, file: &str, overwrite: bool) {
     println!("\nIndex:\n{url}");
     println!("\nStorage files:\n{file}");
@@ -156,7 +102,6 @@ fn print_startup_info(url: &str, file: &str, overwrite: bool) {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tempfile::TempDir;
 
     #[test]
     fn test_init_with_localhost() {
@@ -176,64 +121,5 @@ mod tests {
     fn test_init_with_invalid_host() {
         let result = init("invalid.host", 3000);
         assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_init_storage_file_creates_default_content() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("test.json");
-        let file_path_str = file_path.to_str().unwrap();
-        
-        let result = init_storage_file(file_path_str, false);
-        assert!(result.is_ok());
-        
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert!(content.contains("\"posts\""));
-        assert!(content.contains("\"Hello World\""));
-        assert!(content.contains("\"profile\""));
-        assert!(content.contains("\"Sample User\""));
-    }
-
-    #[test]
-    fn test_init_storage_file_creates_empty_content() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("test.json");
-        let file_path_str = file_path.to_str().unwrap();
-        
-        let result = init_storage_file(file_path_str, true);
-        assert!(result.is_ok());
-        
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert!(content.contains("\"posts\": []"));
-        assert!(content.contains("\"profile\": {}"));
-        assert!(!content.contains("Hello World"));
-        assert!(!content.contains("Sample User"));
-    }
-
-    #[test]
-    fn test_init_storage_file_creates_directories() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("sub").join("dir").join("test.json");
-        let file_path_str = file_path.to_str().unwrap();
-        
-        let result = init_storage_file(file_path_str, false);
-        assert!(result.is_ok());
-        
-        assert!(file_path.exists());
-        let content = fs::read_to_string(&file_path).unwrap();
-        assert!(content.contains("\"posts\""));
-    }
-
-    #[test]
-    fn test_init_storage_file_with_existing_file() {
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("existing.json");
-        let file_path_str = file_path.to_str().unwrap();
-        
-        fs::write(&file_path, "existing content").unwrap();
-        assert!(file_path.exists());
-        
-        let result = init_storage_file(file_path_str, false);
-        assert!(result.is_ok());
     }
 }
